% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testhelper-package.r
\docType{package}
\name{testhelper-package}
\alias{testhelper}
\alias{testhelper-package}
\title{Tools for supporting package testing}
\description{
Currently only a function to determine the correct working
directory when executing unit tests line-by-line.
}
\details{
Details
}
\section{Classes and constructors}{


\itemize{
   \item{\code{\link[testhelper]{testhelper}}: }{

     TODO
   }
}
}

\section{Main functions/methods}{


 \itemize{
   \item{\code{\link[testhelper]{testhelper}}: }{

     TODO
   }
}
}

\section{Disclaimer}{

\strong{This package is work-in-progress and is thus likely to change quite
substantially until it reaches a more mature state.}
}

\section{Aknowledgements}{

This package was inspired by the following resources:

\itemize{
   \item{Package \code{testthat}: } {

     \url{https://github.com/hadley/testthat/}
   }
}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/testhelper}
}
\seealso{
See \code{vignette("introduction", package = "testhelper")}.
}

